sw/requirements.txt
numpy
opencv-python
flake8
Pillow
requests